requirements-dev.txt
# Development Requirements
# ========================
# This file contains development dependencies for the Floorplan integration
#
# Install with: pip install -r requirements-dev.txt
#
# Note: Home Assistant is mocked for testing to avoid full HA installation.
#       For integration tests with real HA, uncomment homeassistant below.

# Testing Framework
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Code Quality & Linting
pylint>=2.17.0
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.0.0

# Type Stubs for type checking
types-PyYAML>=6.0.0
types-requests>=2.28.0

# Data Format Support
PyYAML>=6.0.0

# Optional: Full Home Assistant Integration Testing
# Uncomment the line below if you want to run integration tests with real HA
# homeassistant>=2024.1.0
